+++ bika/lims/upgrade/v01_02_009.py
@@ -8,6 +8,7 @@
 import time
 import transaction
 
+from bika.lims import api
 from bika.lims import logger
 from bika.lims.config import PROJECTNAME as product
 from bika.lims.upgrade import upgradestep
@@ -31,6 +32,7 @@ def upgrade(tool):
     logger.info("Upgrading {0}: {1} -> {2}".format(product, ver_from, version))
 
     # -------- ADD YOUR STUFF HERE --------
+    migrate_attachment_report_options(portal)
 
     # Reindex object security for client contents (see #991)
     reindex_client_local_owner_permissions(portal)
@@ -57,3 +59,21 @@ def reindex_client_local_owner_permissions(portal):
     logger.info("Fixing local owner role on client objects took {:.2f}s"
                 .format(end-start))
     transaction.commit()
+
+    
+def migrate_attachment_report_options(portal):
+    """Migrate Attachments with the report option "a" (attach in report)
+       to the option to "i" (ignore in report)
+    """
+    attachments = api.search({"portal_type": "Attachment"})
+    total = len(attachments)
+    logger.info("Migrating 'Attach to Report' -> 'Ingore in Report' "
+                "for %d attachments" % total)
+    for num, attachment in enumerate(attachments):
+        obj = api.get_object(attachment)
+
+        if obj.getReportOption() in ["a", ""]:
+            obj.setReportOption("i")
+            obj.reindexObject()
+            logger.info("Migrated Attachment %s" % obj.getTextTitle())
+